Actualizacion-Alumno-Profesor

+++ Models/AlumnoMMV/AlumnoModelView.cs
@@ -3,8 +3,8 @@
     public class AlumnoModelView
     {
         public AlumnoModel alumno { get; set; } = new AlumnoModel(); 
-        public int CalContinua { get; set; }
-        public int CalExMedia { get; set; }
-        public int CaliExFinal { get; set; }
+        public float CalContinua { get; set; }
+        public float CalExMedia { get; set; }
+        public float CaliExFinal { get; set; }
     }
 }
